requirements.txt
numpy
scipy
matplotlib
streamlit
fpdf2
